journal: allow access to cleared-conflict TLFs
When a conflict is cleared, and when the journal manager starts up,
the journal manager creates a fake TLF ID for the conflict branch view
of the TLF, and returns that TLF ID for requests that use a handle
specifying a local conflict view.  MDs fetched from that journal have
their TLF ID set to this fake ID, so that upper layers will be none
the wiser.

Issue: KBFS-4138